sc-5491 (4b+4c): wire the candle InstantID worker lane off-Mac
Neutralize the InstantID worker path so it compiles + runs on the Windows/CUDA
candle build, and dispatch InstantID character jobs to it (epic 5480, Phase 3).
4b (neutralize) and 4c (dispatch) land together: the dispatch is what *uses*
generate_instantid_stream, so without it the function would be dead-code under
clippy -D warnings.

The candle `InstantId` API differs from MLX only at the load boundary; the
per-item generate/restore loop is backend-neutral over the SHARED gen_core
contract (reference is `gen_core::Image`, the IdentityNet is `gen_core::
WeightsSource`, the request carries `gen_core::CancelFlag`/`Progress` — all the
exact types `candle_gen_instantid::InstantId::generate` consumes, which only
holds because the sc-4482 skew gate now resolves ONE gen-core rev). So the
divergence is cfg-split in just the load closure:
- with_face: MLX loads two `Weights` files; candle's FaceEmbedder (sc-5490) loads
  the SCRFD/ArcFace pair from THEIR DIRECTORY (the filenames ensure_instantid_
  weights stages), so it takes the dir.
- quantize: MLX-only knob; the candle stack runs dense f16 (quant forced off, not
  recorded).
- largest_face: MLX takes raw RGB bytes + dims; candle takes the `gen_core::Image`.
- the generate* / restore_face calls are `&mut self` on candle (face IP tokens set
  on the UNet per call) vs `&self` on MLX — handled by one `let mut model` rebind.
- the asset/telemetry engine id is `candle_instantid` vs `mlx_instantid`.

Plumbing the candle lane needed:
- un-gate `start_gen_stream` (stream.rs), `mod advanced` (lib.rs), and the
  runtime HF-file download path `ensure_hf_cached_file`/`ensure_cached_file`/
  `remote_content_length` (downloads.rs) to `any(macos, windows+candle)` — the
  InstantID lane is the first off-Mac caller of each.
- relocate the shared `non_empty` + pose/angle-prompt helpers (`pose_entries`,
  `parse_poses`/`PoseInput`, `augment_prompt_for_angle` + its `angle_prompt_
  augment`/`strip_base_prompt`) from the macOS-only `sdxl.rs`/`zimage.rs` into the
  shared `base.rs` (same module, so the macOS callers are unchanged).
- un-gate `include!("image_jobs/instantid.rs")` + add the candle import block.
- dispatch: a Windows/candle branch routes an `instantid_available` job to
  generate_instantid_stream (checked before is_candle_engine, since
  instantid_realvisxl is not an inventory engine id), and the candle `plan` bakes
  the InstantID image count (angle/pose set N) like the macOS route does.

Verified on Windows/CUDA (sm_120): cargo check + clippy -D warnings clean with
--features backend-candle. The off-Mac router half + real-weight GPU validation
follow.
